Added adding new users to the database

+++ db/models/users.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, BigInteger
 from .base import BaseModel
 
 
 class User(BaseModel):
     __tablename__ = 'users'
 
-    telegram_id = Column(Integer(), unique=True)
+    telegram_id = Column(BigInteger(), unique=True)
     full_name = Column(String(300), nullable=True)
     post = Column(Integer(), ForeignKey('posts.id'))
